requirements.txt
fastapi
pydantic-settings
email-validator
uvicorn[standard]
sqlalchemy
databases[aiosqlite]
python-dotenv
rich
asgi-correlation-id
python-json-logger
logtail-python
python-jose
python-multipart
passlib
bcrypt==4.0.1
aiofiles
b2sdk
alembic
pydantic-extra-types
phonenumbers